Add files via upload
The project is a neural network model based on CNN to make the machine identify the image as a human or an animal. The model was trained with several images of respective two classes along with Image Augmentation to avoid overfitting. Overall the model gave good accuracy and did well in validation process